refactored Token class to be public
